Refactor Plugins.Program
